// src/types/puv.ts
export interface PUVFormData {
  colaborador: string;
  gestor: string;
  meus: string; // Produtos ou serviços
  ajudam: string; // Segmento do cliente
  queDesejam: string; // Necessidades do cliente
  para: string; // Verbo próprio (reduzir, evitar)
  dor: string; // Uma dor do cliente
  e: string; // Verbo próprio (aumentar, possibilitar)
  ganho: string; // Um ganho do cliente
}